ap/app/requirements.txt
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.4.4
gunicorn
mysqlclient
oso~=0.9.0
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
SQLAlchemy==1.3.20
SQLAlchemy-bulk-lazy-loader==0.9.9
sqlalchemy-oso~=0.2.1
sqlparse==0.4.1
ulid-py
